// pslpwcl/src/app/modules/pslpapi/model/candidatura.ts
/*
 * ========================LICENSE_START=================================
 * Copyright (C) 2025 Regione Piemonte
 * SPDX-FileCopyrightText: Copyright 2025 | Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AbilitazionePosseduta } from './abilitazionePosseduta';
import { AmbitoDiffusione } from './ambitoDiffusione';
import { BlpLavAnagrafica } from './blpLavAnagrafica';
import { EsperienzaLav } from './esperienzaLav';
import { Fonte } from './fonte';
import { Formazione } from './formazione';
import { InformaticaDich } from './informaticaDich';
import { Intermediario } from './intermediario';
import { IscrizioneAlboDich } from './iscrizioneAlboDich';
import { IstruzioneDich } from './istruzioneDich';
import { LinguaDich } from './linguaDich';
import { ListaSpeciale } from './listaSpeciale';
import { PatentePosseduta } from './patentePosseduta';
import { ProfessioneDesiderata } from './professioneDesiderata';
import { StatoCandidatura } from './statoCandidatura';
import { TipoComunicazione } from './tipoComunicazione';

export interface Candidatura { 
    abilitazionePossedutaList?: Array<AbilitazionePosseduta>;
    codCandidatura?: string;
    codStatoCandidatura?: StatoCandidatura;
    codUserAggiorn?: string;
    codUserInserim?: string;
    dAggiorn?: Date;
    dInserim?: Date;
    dInvio?: Date;
    dScadenza?: Date;
    emailInvio?: string;
    esperienzaLavList?: Array<EsperienzaLav>;
    flgGeneratoDaSistema?: string;
    flgL68?: string;
    flgMezzoProprio?: string;
    flgStampaL68?: string;
    flgTrasferte?: string;
    flgVisibilitaIntermediario?: string;
    formazioneList?: Array<Formazione>;
    id?: number;
    idAmbitoDiffusione?: AmbitoDiffusione;
    idFonteCandidatura: Fonte;
    idIntermediario?: Intermediario;
    idLavAnagrafica: BlpLavAnagrafica;
    idListaSpeciale?: ListaSpeciale;
    idTipoComunicazione?: TipoComunicazione;
    incentiviEQualita?: string;
    informaticaDichList?: Array<InformaticaDich>;
    iscrizioneAlboDichList?: Array<IscrizioneAlboDich>;
    istruzioneDichList?: Array<IstruzioneDich>;
    linguaDichList?: Array<LinguaDich>;
    paroleChiaveMatching?: string;
    patentePossedutaList?: Array<PatentePosseduta>;
    professioneDesiderataList?: Array<ProfessioneDesiderata>;
    titoloCv: string;
    ulterioriInformazioni?: string;
}